Move system

+++ DumbKnights/components/RenderingSystem.cpp
@@ -5,6 +5,7 @@
 #include "RenderingSystem.h"
 #include "../Locator.h"
 #include <Texture.h>
+#include "Components.h"
 
 void RenderingSystem::draw(engine::ecs::Coordinator& coordinator, engine::render::IRenderer& renderer) {
     for (auto& entity: entities) {
@@ -24,7 +25,8 @@ void RenderingSystem::draw(engine::ecs::Coordinator& coordinator, engine::render
         auto texture = Locator::instance().assets().getTexture(sprite.textureName).get();
         gmath::Rectangle dstRect{
                 Vec2{transform.position.x + sprite.origin.x, transform.position.y + sprite.origin.y},
-                Vec2{texture->width * transform.scale.x, texture->height * transform.scale.y}
+                Vec2{static_cast<f32>(texture->width) * transform.scale.x,
+                     static_cast<f32>(texture->height) * transform.scale.y}
         };
         renderer.drawSprite(texture, gmath::Rectangle::nullRectangle, dstRect, transform.rotation,
                             sprite.origin, sprite.flip);
